requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
piper-tts==1.2.0
httpx==0.27.2
readability-lxml==0.8.1
lxml-html-clean==0.4.1
